# backend/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "tittle": "Weather Data API",
    "description": "JSON REST API for tracking data of the weather sightings",
    "title": "REST API"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/tests/{id}": {
      "get": {
        "summary": "Get a specific test by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific test by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/tests/": {
      "post": {
        "summary": "Create a specific test",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new test.",
          "content": {
            "application/json": {
              "schema": {
                "country": "string",
                "city": "string"
              },
              "example": {
                "country": "any",
                "city": "any"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update a specific test by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new test",
          "content": {
            "application/json": {
              "schema": {
                "id": {
                  "type": "string"
                },
                "country": {
                  "type": "string",
                  "pattern": "^[A-Za-z ]+$",
                  "maxLength": 50
                },
                "city": {
                  "type": "string",
                  "pattern": "^[A-Za-z ]+$",
                  "maxLength": 50
                }
              },
              "example": {
                "id": 1,
                "country": "Colombia"
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific test by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new test",
          "content": {
            "application/json": {
              "schema": {
                "id": {
                  "type": "string"
                },
                "country": {
                  "type": "string",
                  "pattern": "^[A-Za-z ]+$",
                  "maxLength": 50
                },
                "city": {
                  "type": "string",
                  "pattern": "^[A-Za-z ]+$",
                  "maxLength": 50
                }
              },
              "example": {
                "id": 1,
                "country": "Colombia",
                "city": "Bogota"
              }
            }
          }
        }
      }
    },
    "/tests/many": {
      "post": {
        "summary": "Create a many Tests",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Adding new Test",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "country": {
                      "type": "string"
                    },
                    "city": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "authenticationKey": "829ef690-a373-4485-9afb-b9f4b8f80f9a",
                    "test": [
                      {
                        "country": "Tests for create many 1",
                        "city": "Tests for create many 1"
                      },
                      {
                        "country": "Tests for create many 2",
                        "city": "Tests for create many 2"
                      },
                      {
                        "country": "Tests for create many 3",
                        "city": "Tests for create many 3"
                      },
                      {
                        "country": "Tests for create many 4",
                        "city": "Tests for create many 4"
                      },
                      {
                        "country": "Tests for create many 5",
                        "city": "Tests for create many 5"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tests/deleteMany": {
      "delete": {
        "summary": "Delete many tests by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Deleting many tests",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "authenticationKey": "b2317874-ab1c-4d71-9883-2bba10b71ebd",
                "ids": [
                  "64230e5570a4fd40cb4d2ca9",
                  "6423a3be27a38a7885ecdbc6"
                ]
              }
            }
          }
        }
      }
    }
  }
}